feat(emulators): move installation into emulators
