+++ core/package.json
@@ -16,7 +16,6 @@
   "dependencies": {
     "@secret-agent/commons": "1.4.1-alpha.4",
     "@secret-agent/interfaces": "1.4.1-alpha.4",
-    "@secret-agent/emulate-browsers-base": "1.4.1-alpha.4",
     "@secret-agent/emulate-chrome-latest": "1.4.1-alpha.4",
     "@secret-agent/emulate-humans-basic": "1.4.1-alpha.4",
     "@secret-agent/emulate-humans-ghost": "1.4.1-alpha.4",
@@ -35,6 +34,7 @@
     "@secret-agent/client": "1.4.1-alpha.4",
     "@secret-agent/core": "1.4.1-alpha.4",
     "@secret-agent/puppet-chrome": "1.4.1-alpha.4",
+    "@secret-agent/emulate-browsers-installer": "1.4.1-alpha.4",
     "@secret-agent/testing": "1.4.1-alpha.4",
     "@types/better-sqlite3": "^5.4.1",
     "@types/json-socket": "^0.1.17",
