+++ emulate-browsers/installer/lib/DependencyInstaller.ts
