+++ emulate-browsers/installer/lib/LinuxUtils.ts
