+++ emulate-browsers/installer/lib/validateHostDependencies.ts
