+++ emulate-browsers/installer/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "@secret-agent/emulate-browsers-installer",
+  "version": "1.4.1-alpha.4",
+  "description": "SecretAgent BrowserEmulator installer",
+  "bin": {
+    "install-browser-deps": "./bin/installDeps.js"
+  },
+  "scripts": {},
+  "dependencies": {
+    "@secret-agent/commons": "1.4.1-alpha.4",
+    "@secret-agent/interfaces": "1.4.1-alpha.4",
+    "progress": "^2.0.3",
+    "tar": "^6.1.0"
+  }
+}
