+++ interfaces/IBrowserEngine.ts
@@ -6,6 +6,7 @@ export default interface IBrowserEngine {
   executablePath: string;
   executablePathEnvVar: string;
   isHeaded?: boolean;
+  verifyLaunchable?(): Promise<any>;
   getLaunchArguments?(puppetOptions: IPuppetLaunchArgs, defaultArguments: string[]): string[];
 }
