+++ puppet/package.json
@@ -2,15 +2,10 @@
   "name": "@secret-agent/puppet",
   "version": "1.4.1-alpha.4",
   "description": "Puppet driver",
-  "bin": {
-    "puppet-install-deps": "./bin/installDeps.js"
-  },
   "dependencies": {
     "@secret-agent/commons": "1.4.1-alpha.4",
     "@secret-agent/interfaces": "1.4.1-alpha.4",
     "@secret-agent/puppet-chrome": "1.4.1-alpha.4",
-    "progress": "^2.0.3",
-    "tar": "^6.1.0",
     "ws": "^7.4.4"
   },
   "devDependencies": {
